Refactor project structure, remove dead code and add docs
-  Especially in _base
-  Establish `inputHandling` dir for stuff related to.. input validation, sanitization, extraction from commander options etc
-  Rename 'cli' to 'terminal' and keep only stuff that related to fiddling with terminal IO: logging, clearing & prompting using inquirer
-  Move input validation error to be a top-level type. input validation errors may fly from different places as well

RELATED: RAIL-3428
