+++ tools/plugin-toolkit/src/_base/terminal/clear.ts
