+++ tools/plugin-toolkit/src/addPluginCmd/index.ts
@@ -1,8 +1,7 @@
 // (C) 2021 GoodData Corporation
-import { ActionOptions } from "../_base/types";
-import { logError, logInfo, logSuccess, logWarn } from "../_base/cli/loggers";
+import { ActionOptions, isInputValidationError } from "../_base/types";
+import { logError, logInfo, logSuccess, logWarn } from "../_base/terminal/loggers";
 import { AddCmdActionConfig, getAddCmdActionConfig } from "./actionConfig";
-import { isInputValidationError } from "../_base/cli/validators";
 import fse from "fs-extra";
 import { IDashboardPlugin, isNotAuthenticated } from "@gooddata/sdk-backend-spi";
