+++ tools/plugin-toolkit/src/usePluginCmd/index.ts
@@ -1,6 +1,6 @@
 // (C) 2021 GoodData Corporation
 import { ActionOptions } from "../_base/types";
-import { logInfo } from "../_base/cli/loggers";
+import { logInfo } from "../_base/terminal/loggers";
 
 export async function usePluginCmdAction(options: ActionOptions): Promise<void> {
     logInfo(`usePluginCmdAction ${JSON.stringify(options, null, 4)}`);
